Removed special IO mode, validate, x11 module, ...
The X11 module has been improved but still needs a lot of work. Lines can be
drawn, windows opened, lines erased, events detected.

The validate function now detects closed objects, the validate subroutine
still needs a lot of work doing to it.

A few new macros and a function of little consequence have been added to
the "liblisp.h" header. This header needs going through and the functions
renaming for consistency.

+++ mod/liblisp_tcc.c
@@ -8,6 +8,7 @@
 #include <libtcc.h>
 #include <liblisp.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 extern lisp *lglobal; /* from main.c */
 
@@ -35,7 +36,7 @@ static struct module_subroutines { subr p; char *name; } primitives[] = {
 };
 #undef X
 
-static int ud_tcc = 0;
+static intptr_t ud_tcc = 0;
 
 static void ud_tcc_free(cell *f) {
         tcc_delete(userval(f));
